+++ src/common/types/index.ts
@@ -2,4 +2,5 @@ export * from './Stoppable.js';
 export * from './Resettable.js';
 export * from './Resizable.js';
 export * from './Updatable.js';
+export * from './World.js';
 export * from './Prettify.js';